requirements.txt
opencv-python-headless
numpy
matplotlib
streamlit
vision-agent
reportlab
pillow
python-dotenv
